+++ src/main/java/io/cdap/plugin/dynamo/source/DynamoDBBatchSource.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package io.cdap.plugin.batch;
+package io.cdap.plugin.dynamo.source;
 
 import com.amazonaws.services.dynamodbv2.document.Item;
 import com.google.common.base.Strings;
