+++ src/main/java/io/cdap/plugin/dynamo/source/DynamoDBRecordReader.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package io.cdap.plugin.batch;
+package io.cdap.plugin.dynamo.source;
 
 import com.amazonaws.auth.BasicAWSCredentials;
 import com.amazonaws.regions.RegionUtils;
