+++ src/main/java/io/cdap/plugin/dynamo/source/split/DynamoDBSegmentsSplit.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package io.cdap.plugin.batch.split;
+package io.cdap.plugin.dynamo.source.split;
 
 import org.apache.hadoop.io.Writable;
 import org.apache.hadoop.mapreduce.InputSplit;
